% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berry.R
\encoding{UTF-8}
\name{get_berry}
\alias{get_berry}
\alias{get_berry_firmness}
\alias{get_berry_flavour}
\alias{get_berry_flavor}
\title{Get Berry Information}
\usage{
get_berry(id)

get_berry_firmness(id)

get_berry_flavour(id)

get_berry_flavor(id)
}
\arguments{
\item{id}{Either the numeric ID of the required item from the category,
or the character string of the item}
}
\value{
A named list of all related data to the requested entity
}
\description{
A series of functions that pull berry data from the PokéAPI
}
\details{
If you type the wrong name and it's close enough to one of the available
names, then the error message will give you the correct spelling.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Character name
get_berry("rawst")

# ID
get_berry(1)
\dontshow{\}) # examplesIf}
}
\seealso{
\url{https://pokeapi.co/docs/v2#berries-section}
}
