% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\encoding{UTF-8}
\name{call_pk_api}
\alias{call_pk_api}
\title{API Call Wrapper}
\usage{
call_pk_api(category, id = NULL, subcategory = NULL)
}
\arguments{
\item{category}{Name of the category endpoint to pull data from}

\item{id}{Either the numeric ID or the name of the item required from `category`}

\item{subcategory}{For the Pokémon category, there are some subcategories like
`"encounter"`. Most endpoints do not have one, so the default is `NULL`}
}
\value{
A named list of all related data to the requested category and ID
}
\description{
A basic wrapper that can handle all category and ID queries from the
PokéAPI.
}
\details{
This function also handles looking at the cache to see whether the
data has already been pulled and can easily be loaded rather than
call the API each time.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Pokemon
call_pk_api("pokemon", "bulbasaur")

# Generation 7
call_pk_api("generation", 7)
\dontshow{\}) # examplesIf}
}
