% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contest.R
\encoding{UTF-8}
\name{get_contest_type}
\alias{get_contest_type}
\alias{get_contest_effect}
\alias{get_super_contest_effect}
\title{Get Contest Information}
\usage{
get_contest_type(id)

get_contest_effect(id)

get_super_contest_effect(id)
}
\arguments{
\item{id}{Either the numeric ID of the required item from the category,
or the character string of the item}
}
\value{
A named list of all related data to the requested entity
}
\description{
A series of functions that pull contset data from the PokéAPI
}
\details{
`get_contest_effect` and `get_super_contest_effect` only take a numeric ID
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Character name
get_contest_type("cool")

# ID
get_contest_type(2)
\dontshow{\}) # examplesIf}
}
\seealso{
\url{https://pokeapi.co/docs/v2#contests-section}
}
