% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encounter.R
\name{get_encounter_method}
\alias{get_encounter_method}
\alias{get_encounter_condition}
\alias{get_encounter_condition_value}
\title{Get Encounter Information}
\usage{
get_encounter_method(id)

get_encounter_condition(id)

get_encounter_condition_value(id)
}
\arguments{
\item{id}{Either the numeric ID of the required item from the category,
or the character string of the item}
}
\value{
A named list of all related data to the requested entity
}
\description{
A series of functions that pull encounter data from the PokéAPI
}
\details{
If you type the wrong name and it's close enough to one of the available
names, then the error message will give you the correct spelling.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Character name
get_encounter_method("old-rod")

# ID
get_encounter_method(1)
\dontshow{\}) # examplesIf}
}
\seealso{
\url{https://pokeapi.co/docs/v2#encounters-section}
}
