% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution.R
\name{get_evolution_chain}
\alias{get_evolution_chain}
\alias{get_evolution_trigger}
\title{Get Evolution Information}
\usage{
get_evolution_chain(id)

get_evolution_trigger(id)
}
\arguments{
\item{id}{Either the numeric ID of the required item from the category,
or the character string of the item}
}
\value{
A named list of all related data to the requested entity
}
\description{
A series of functions that pull evolution data from the PokéAPI
}
\details{
`get_evolution_chain` requires the Pokémon ID rather than the
name of the revelant Pokémon
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Character name
get_evolution_trigger("use-item")

# ID
get_evolution_chain(1)
\dontshow{\}) # examplesIf}
}
\seealso{
\url{https://pokeapi.co/docs/v2#evolutions-section}
}
