% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\encoding{UTF-8}
\name{get_generation}
\alias{get_generation}
\alias{get_pokedex}
\alias{get_version}
\alias{get_version_groups}
\title{Get Games Information}
\usage{
get_generation(id)

get_pokedex(id)

get_version(id)

get_version_groups(id)
}
\arguments{
\item{id}{Either the numeric ID of the required item from the category,
or the character string of the item}
}
\value{
A named list of all related data to the requested entity
}
\description{
A series of functions that pull Pokémon games data from the PokéAPI
}
\details{
If you type the wrong name and it's close enough to one of the available
names, then the error message will give you the correct spelling.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Character name
get_generation("generation-iii")

# ID
get_generation(3)
\dontshow{\}) # examplesIf}
}
\seealso{
\url{https://pokeapi.co/docs/v2#games-section}
}
