% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.R
\encoding{UTF-8}
\name{get_item}
\alias{get_item}
\alias{get_item_attribute}
\alias{get_item_category}
\alias{get_item_fling_effect}
\alias{get_item_pocket}
\title{Get Items Information}
\usage{
get_item(id)

get_item_attribute(id)

get_item_category(id)

get_item_fling_effect(id)

get_item_pocket(id)
}
\arguments{
\item{id}{Either the numeric ID of the required item from the category,
or the character string of the item}
}
\value{
A named list of all related data to the requested entity
}
\description{
A series of functions that pull Pokémon items data from the PokéAPI
}
\details{
If you type the wrong name and it's close enough to one of the available
names, then the error message will give you the correct spelling.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Character name
get_item("old-rod")

# ID
get_item(1)
\dontshow{\}) # examplesIf}
}
\seealso{
\url{https://pokeapi.co/docs/v2#items-section}
}
