% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\encoding{UTF-8}
\name{get_location}
\alias{get_location}
\alias{get_location_area}
\alias{get_pal_park_area}
\title{Get Location Information}
\usage{
get_location(id)

get_location_area(id)

get_pal_park_area(id)
}
\arguments{
\item{id}{Either the numeric ID of the required item from the category,
or the character string of the item}
}
\value{
A named list of all related data to the requested entity
}
\description{
A series of functions that pull Pokémon locations data from the PokéAPI
}
\details{
If you type the wrong name and it's close enough to one of the available
names, then the error message will give you the correct spelling.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Character name
get_location("celadon-city")

# ID
get_location(67)
\dontshow{\}) # examplesIf}
}
\seealso{
\url{https://pokeapi.co/docs/v2#locations-section}
}
