% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine.R
\encoding{UTF-8}
\name{get_machine}
\alias{get_machine}
\title{Get TM/HM Information}
\usage{
get_machine(id)
}
\arguments{
\item{id}{Either the numeric ID of the required item from the category,
or the character string of the item}
}
\value{
A named list of all related data to the requested entity
}
\description{
A series of functions that pull Pokémon machine data from the PokéAPI
}
\details{
`get_machine` only takes a numeric ID
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_machine(67)
\dontshow{\}) # examplesIf}
}
\seealso{
\url{https://pokeapi.co/docs/v2#machines-section}
}
