% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeapi-package.R
\docType{package}
\name{pokeapi-package}
\alias{pokeapi}
\alias{pokeapi-package}
\title{pokeapi: R wrapper for 'PokéAPI'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

'PokéAPI' is a lightweight 'API' that allows the user to pull almost any information about 'Pokémon' one can think of. 'pokeapi' provides a wrapper for this 'API'.
}
\author{
\strong{Maintainer}: Ashley Baldry \email{arbaldry91@gmail.com}

}
