% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clear_pk_cache}
\alias{clear_pk_cache}
\title{PokéAPI Cache}
\usage{
clear_pk_cache(version = POKEAPI_VERSION)
}
\arguments{
\item{version}{Version of PokéAPI to use. Default is set to current
version available.}
}
\value{
If a cache exists then `clear_pokeapi_cache` will delete all previous data
pulls from the PokéAPI.
}
\description{
To reduce the amount of API calls, successful results will be stored
in a cache and can be retrieved with subsequent calls to the API.
}
\details{
If you want to call the API each time, set the environment
variable `R_POKEAPI_CACHE = FALSE`
}
