% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeinfo.R
\name{get_pokemon}
\alias{get_pokemon}
\alias{get_pokemon_ability}
\alias{get_pokemon_characteristic}
\alias{get_egg_group}
\alias{get_pokemon_gender}
\title{Get Pokéinfo}
\usage{
get_pokemon(id)

get_pokemon_ability(id)

get_pokemon_characteristic(id)

get_egg_group(id)

get_pokemon_gender(id)
}
\arguments{
\item{id}{Either the numeric ID of the required item from the category,
or the character string of the item}
}
\value{
A named list of all related data to the requested entity
}
\description{
A series of functions that pull data from the PokéAPI
}
\details{
If you type the wrong name and it's close enough to one of the available
names, then the error message will give you the correct spelling.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Character name
get_pokemon("bulbasaur")

# ID
get_pokemon(1)
\dontshow{\}) # examplesIf}
}
\seealso{
\url{https://pokeapi.co/docs/v2}
}
