% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\encoding{UTF-8}
\name{get_pokemon}
\alias{get_pokemon}
\alias{get_ability}
\alias{get_characteristic}
\alias{get_egg_group}
\alias{get_gender}
\alias{get_growth_rate}
\alias{get_nature}
\alias{get_pokeathlon_stat}
\alias{get_pokemon_location_areas}
\alias{get_pokemon_colour}
\alias{get_pokemon_color}
\alias{get_pokemon_form}
\alias{get_pokemon_habitat}
\alias{get_pokemon_shape}
\alias{get_pokemon_species}
\alias{get_stat}
\alias{get_type}
\title{Get Pokémon Information}
\usage{
get_pokemon(id)

get_ability(id)

get_characteristic(id)

get_egg_group(id)

get_gender(id)

get_growth_rate(id)

get_nature(id)

get_pokeathlon_stat(id)

get_pokemon_location_areas(id)

get_pokemon_colour(id)

get_pokemon_color(id)

get_pokemon_form(id)

get_pokemon_habitat(id)

get_pokemon_shape(id)

get_pokemon_species(id)

get_stat(id)

get_type(id)
}
\arguments{
\item{id}{Either the numeric ID of the required item from the category,
or the character string of the item}
}
\value{
A named list of all related data to the requested entity
}
\description{
A series of functions that pull data from the PokéAPI
}
\details{
If you type the wrong name and it's close enough to one of the available
names, then the error message will give you the correct spelling.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Character name
get_pokemon("bulbasaur")

# ID
get_pokemon(1)
\dontshow{\}) # examplesIf}
}
\seealso{
\url{https://pokeapi.co/docs/v2#pokemon-section}
}
