% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\encoding{UTF-8}
\name{random_pk_id}
\alias{random_pk_id}
\title{Get Random Category ID}
\usage{
random_pk_id(category)
}
\arguments{
\item{category}{Name of the category endpoint to pull data from}
}
\description{
To make random teams/moves, use this to get a random ID to pull
from the PokéAPI.
}
\details{
This uses internal data, so may not include the latest information.
}
