% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeapi-package.R
\docType{package}
\name{pokeapi-package}
\alias{pokeapi}
\alias{pokeapi-package}
\title{pokeapi: R wrapper for 'PokéAPI'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Ashley Baldry \email{arbaldry91@gmail.com}

}
